+++ tools/markdown-generator/CHANGELOG.md
@@ -1,5 +1,13 @@
 # markdown-generator
 
+## 0.1.51
+
+### Patch Changes
+
+- Updated dependencies [[`5684c8a`](https://github.com/adobe/spectrum-design-data/commit/5684c8afbf34fc391717fdb975699e06eb677a4e), [`018f18e`](https://github.com/adobe/spectrum-design-data/commit/018f18e4175e4058fee2ad8f7678c8b07bd278c8)]:
+  - @adobe/spectrum-design-data@0.16.0
+  - @adobe/spectrum-component-api-schemas@6.1.34
+
 ## 0.1.50
 
 ### Patch Changes