+++ tools/spectrum-design-data-mcp/CHANGELOG.md
@@ -1,5 +1,12 @@
 # [**@adobe/spectrum-design-data-mcp**](https://github.com/adobe/spectrum-design-data-mcp)
 
+## 1.1.49
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @adobe/spectrum-component-api-schemas@6.1.34
+
 ## 1.1.48
 
 ### Patch Changes